exclude TrgUnit.ignoreStrategicSuicideMissions
flag 1 at 0x660178 seems to be completely unrelated to strategic
suicide mission. it's only checked with units that have already been
added to an attack wave, and looks like it prevents those units from
being counted as being missing from the attack group or something
like that. i'm not fully understanding its purpose yet and not seeing
much of a noticeable effect by ticking it on for regular military units.